Move TranscriptFormat to separate file
