+++ Modules/DataModel/Sources/PocketCastsDataModel/Public/Model/Episode.swift
@@ -205,44 +205,7 @@ public class Episode: NSObject, BaseEpisode {
         public struct Transcript: Decodable {
             public let url: String
             public let type: String
-            public let language: String?
-
-            public var transcriptFormat: TranscriptFormat? {
-                return TranscriptFormat(rawValue: self.type)
-            }
-        }
-
-        public enum TranscriptFormat: String {
-
-            case srt = "application/srt"
-            case vtt = "text/vtt"
-            case textHTML = "text/html"
-            case jsonPodcastIndex = "application/json"
-
-            public var fileExtension: String {
-                switch self {
-                case .srt:
-                    return "srt"
-                case .vtt:
-                    return "vtt"
-                case .textHTML:
-                    return "html"
-                case .jsonPodcastIndex:
-                    return "json"
-                }
-            }
-
-            // Transcript formats we support in order of priority of use
-            public static let supportedFormats: [TranscriptFormat] = [.jsonPodcastIndex, .vtt, .srt, .textHTML]
-
-            public static func bestTranscript(from available: [Transcript]) -> Transcript? {
-                for format in Self.supportedFormats {
-                    if let transcript = available.first(where: { $0.type == format.rawValue}) {
-                        return transcript
-                    }
-                }
-                return available.first
-            }
-        }
+            public let language: String?            
+        }        
     }
 }
